requirements.txt
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gdown==5.2.0
numpy==1.26.4
pillow==11.1.0
gunicorn==20.1.0